# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: Vending Machine API
  description: |
    A comprehensive RESTful API for managing a vending machine system.
    
    ## Features
    - User authentication and authorization
    - Product management (CRUD operations)
    - Purchase processing and transaction tracking
    - Real-time inventory management
    - Sales analytics and reporting
    
    ## Authentication
    This API uses Bearer Token authentication. Obtain a token by logging in via the `/auth/login` endpoint.
    
    ## Rate Limiting
    API requests are limited to 60 requests per minute per IP address.
    
    ## Base URL
    `http://localhost:8000/api/v1`
  version: 1.0.0
  contact:
    name: Vending Machine Support
    email: support@vending-machine.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.vending-machine.com/v1
    description: Production server

paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              successful_login:
                summary: Successful login example
                value:
                  email: "admin@example.com"
                  password: "password"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                successful_response:
                  summary: Successful login response
                  value:
                    status: "success"
                    message: "Login successful"
                    data:
                      user:
                        id: 1
                        name: "Admin User"
                        email: "admin@example.com"
                        role: "admin"
                        created_at: "2024-01-01T00:00:00.000000Z"
                        updated_at: "2024-01-01T00:00:00.000000Z"
                      token: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                      token_type: "bearer"
                      expires_in: 3600
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate the current access token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: "success"
                message: "Logout successful"
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get information about the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /products:
    get:
      tags:
        - Products
      summary: List products
      description: Get a paginated list of products with optional filtering and sorting
      parameters:
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/AvailableOnly'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDirection'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

    post:
      tags:
        - Products
      summary: Create product
      description: Create a new product (requires admin privileges)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            examples:
              create_product:
                summary: Create new product example
                value:
                  name: "Coca Cola"
                  description: "Refreshing cola drink"
                  price: 3.990
                  quantity_available: 50
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'

  /products/{product}:
    get:
      tags:
        - Products
      summary: Get product
      description: Get detailed information about a specific product
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Products
      summary: Update product
      description: Update product information (requires admin privileges)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Products
      summary: Delete product
      description: Delete a product (requires admin privileges)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete product with existing transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: "error"
                message: "Cannot delete product. This product has 5 transaction(s) associated with it."

  /products/{product}/purchase:
    post:
      tags:
        - Products
      summary: Purchase product
      description: Process a product purchase transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
            examples:
              purchase_product:
                summary: Purchase product example
                value:
                  quantity: 2
                  amount_paid: 10.000
      responses:
        '200':
          description: Purchase processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  /products/statistics:
    get:
      tags:
        - Products
      summary: Product statistics
      description: Get product and sales statistics
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from login endpoint"

  parameters:
    ProductId:
      in: path
      name: product
      required: true
      schema:
        type: integer
      description: Product ID
      example: 1

    Search:
      in: query
      name: search
      schema:
        type: string
      description: Search products by name
      example: "Cola"

    AvailableOnly:
      in: query
      name: available_only
      schema:
        type: boolean
        default: false
      description: Filter only available products (quantity > 0)

    SortBy:
      in: query
      name: sort_by
      schema:
        type: string
        enum: [name, price, quantity, created_at]
        default: created_at
      description: Sort field

    SortDirection:
      in: query
      name: direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort direction

    PerPage:
      in: query
      name: per_page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page

    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "admin@example.com"
        password:
          type: string
          minLength: 6
          description: User password
          example: "password"

    LoginResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Login successful"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
            token_type:
              type: string
              example: "bearer"
            expires_in:
              type: integer
              example: 3600

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Admin User"
        email:
          type: string
          format: email
          example: "admin@example.com"
        role:
          type: string
          enum: [admin, user]
          example: "admin"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    UserResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        data:
          $ref: '#/components/schemas/User'

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Coca Cola"
        description:
          type: string
          example: "Refreshing cola drink"
        price:
          type: number
          format: decimal
          example: 3.990
        quantity_available:
          type: integer
          example: 50
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        data:
          $ref: '#/components/schemas/Product'

    ProductListResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            current_page:
              type: integer
              example: 1
            data:
              type: array
              items:
                $ref: '#/components/schemas/Product'
            first_page_url:
              type: string
            from:
              type: integer
            last_page:
              type: integer
            last_page_url:
              type: string
            links:
              type: array
              items:
                type: object
            next_page_url:
              type: string
            path:
              type: string
            per_page:
              type: integer
            prev_page_url:
              type: string
            to:
              type: integer
            total:
              type: integer

    CreateProductRequest:
      type: object
      required:
        - name
        - price
        - quantity_available
      properties:
        name:
          type: string
          maxLength: 255
          example: "Coca Cola"
        description:
          type: string
          nullable: true
          example: "Refreshing cola drink"
        price:
          type: number
          format: decimal
          minimum: 0.001
          maximum: 999999.999
          example: 3.990
        quantity_available:
          type: integer
          minimum: 0
          maximum: 999999
          example: 50

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: "Coca Cola"
        description:
          type: string
          nullable: true
          example: "Refreshing cola drink"
        price:
          type: number
          format: decimal
          minimum: 0.001
          maximum: 999999.999
          example: 3.990
        quantity_available:
          type: integer
          minimum: 0
          maximum: 999999
          example: 50

    PurchaseRequest:
      type: object
      required:
        - quantity
        - amount_paid
      properties:
        quantity:
          type: integer
          minimum: 1
          description: Quantity to purchase
          example: 2
        amount_paid:
          type: number
          format: decimal
          minimum: 0.001
          description: Amount paid by customer
          example: 10.000

    PurchaseResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Purchase successful"
        data:
          type: object
          properties:
            transaction:
              $ref: '#/components/schemas/Transaction'
            change_given:
              type: number
              format: decimal
              example: 2.020

    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 1
        unit_price:
          type: number
          format: decimal
          example: 3.990
        quantity:
          type: integer
          example: 2
        total_price:
          type: number
          format: decimal
          example: 7.980
        amount_paid:
          type: number
          format: decimal
          example: 10.000
        change_given:
          type: number
          format: decimal
          example: 2.020
        status:
          type: string
          enum: [completed, pending, failed]
          example: "completed"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    StatisticsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total_products:
              type: integer
              example: 25
            available_products:
              type: integer
              example: 20
            out_of_stock_products:
              type: integer
              example: 5
            total_transactions:
              type: integer
              example: 150
            today_revenue:
              type: number
              format: decimal
              example: 450.750
            total_revenue:
              type: number
              format: decimal
              example: 12500.000
            average_transaction_value:
              type: number
              format: decimal
              example: 83.333
            top_selling_products:
              type: array
              items:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
                  total_sold:
                    type: integer
                    example: 45
                  revenue:
                    type: number
                    format: decimal
                    example: 179.550

    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "An error occurred"

    ValidationError:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Validation failed"
        errors:
          type: object
          description: Validation errors by field
          example:
            email: ["The email field is required."]
            password: ["The password must be at least 6 characters."]

  responses:
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Unauthorized"

    Forbidden:
      description: Access forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Access forbidden"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Resource not found"

    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

tags:
  - name: Authentication
    description: User authentication and authorization endpoints
  - name: Products
    description: Product management and purchase endpoints